HMAC gadget instances implemented.
